+++ src/disposables.ts
@@ -1,0 +1,3 @@
+import type {Disposable} from 'vscode';
+
+export const disposables = new Set<Disposable>();
