+++ src/output.ts
@@ -1,0 +1,5 @@
+import vscode from 'vscode';
+import {disposables} from './disposables.js';
+
+export const output = vscode.window.createOutputChannel('Fish Completion');
+disposables.add(output);
